Split Option classes into a separate file
This reduces the size of the PControllerPhys script from around 480
lines of code to 385 lines of code.
