+++ Assets/Scripts/Movement/PlayerMovementOptions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 67e5e5ecdd7f2039a9c679d5fc1b8e7a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
